refactor(request-context): split metadata primitives
