+++ crates/bitnet-request-context-core/src/id.rs
@@ -1,0 +1,26 @@
+use std::fmt;
+
+/// Unique request identifier.
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct RequestId(pub String);
+
+impl RequestId {
+    pub fn new(id: impl Into<String>) -> Self {
+        Self(id.into())
+    }
+
+    /// Generate a simple sequential ID.
+    pub fn sequential(n: u64) -> Self {
+        Self(format!("req-{n}"))
+    }
+
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+}
+
+impl fmt::Display for RequestId {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
